requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python
numpy
easyocr
Pillow
